Basics for piFpaF (no smoke)

+++ OpenGL/src/Basics/Coords.h
@@ -14,7 +14,7 @@ struct Coords
 public:
 	double posX;
 	double posY;
-	float Velocity;
+	double Velocity;
 };
 
 #endif
